requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
pillow==10.3.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
django-rest-passwordreset==1.4.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
aiogram~=2.9
aiohttp~=3.9.5
aiosqlite~=0.20.0
Scrapy==2.11.2
Twisted==24.3.0
aiocron==1.8
